extract apply correction
