+++ model/common/tests/interpolation_tests/test_interpolation_fields.py
@@ -231,14 +231,15 @@ def test_compute_c_bln_avg(grid_savepoint, interpolation_savepoint, icon_grid, e
     c_bln_avg_v1 = np.copy(c_bln_avg)
     c_bln_avg_v2 = np.copy(c_bln_avg)
     start = time.time()
-    c_bln_avg_v1 = compute_force_mass_conservation_to_c_bln_avg(c_bln_avg_v1, cell_areas, c2e2c, cell_owner_mask,
-                                                                divavg_cntrwgt, horizontal_start,
+    c2e2c0 = icon_grid.connectivities[dims.C2E2CODim]
+    c_bln_avg_v1 = compute_force_mass_conservation_to_c_bln_avg(c_bln_avg_v1, cell_areas, c2e2c,c2e2c0, cell_owner_mask,
+                                                                divavg_cntrwgt,
                                                                 horizontal_start_p2,
                                                                 niter=iterations)
     d1 = time.time()-start
     print(f" v1: time {d1}")
 
-    c2e2c0 = icon_grid.connectivities[dims.C2E2CODim]
+
     cell_owner_mask = grid_savepoint.c_owner_mask()
     start = time.time()
     # unforced, c_bln_avg_v2 = force_mass_conservation(
